GitHub: Refactor and spread GitHub modules along the feature axis
- Naming things
- Cleanups
