+++ docs/guide/github.md
@@ -4,9 +4,9 @@ Tap into the GitHub API and generate reports in Markdown format.
 
 ## Features
 
-- Report about user activity in [PPP] format.
-- Report about CI failures on [GHA].
-- Report about bugs and similar important items.
+- Actions: Report about CI failures on [GHA].
+- Activity: Report about user activity in [PPP] format.
+- Attention: Report about bugs and similar important items.
 
 ## Setup
 
@@ -29,24 +29,24 @@ export GH_TOKEN=ghp_600VEZtdzinvalid7K2R86JTiKJAAp1wNwVP
 Note that many options are optional. Just omit them in order to expand the
 search scope.
 
-### PPP reports
-Report about activities of individual authors.
-```shell
-rapporto gh ppp --organization=python --author=AA-Turner --timerange="2025-01-01..2025-01-31"
-rapporto gh ppp --organization=python --author=AA-Turner --timerange="2025W04"
-```
-
-### CI reports
+### Actions report
 Report about activities of GitHub Actions workflow runs, mostly failing ones.
 ```shell
-rapporto gh ci --repository=acme/acme-examples
-rapporto gh ci --repositories-file=acme-repositories.txt
+rapporto gh actions --repository=acme/acme-examples
+rapporto gh actions --repositories-file=acme-repositories.txt
 ```
 
-### Importance reports
+### Activity report
+Report about activities of individual authors.
+```shell
+rapporto gh activity --organization=python --author=AA-Turner --timerange="2025-01-01..2025-01-31"
+rapporto gh activity --organization=python --author=AA-Turner --timerange="2025W04"
+```
+
+### Attention report
 Report about important items that deserve your attention, bugs first.
 ```shell
-rapporto gh att --organization=python --timerange="2025W07"
+rapporto gh attention --organization=python --timerange="2025W07"
 ```
 If you want to explore your personal repositories, please use the
 `--organization` option with your username, e.g. `--organization=AA-Turner`.
